user password gettting updated with token

+++ controllers/auth/resetPassword.js
@@ -1,39 +1,62 @@
 const Joi = require("joi");
 const bcrypt = require("bcrypt");
+const JWT = require("jsonwebtoken");
 const {
   fetchUserFromId,
   updatePasswordService,
 } = require("../../services/User");
 const {
-  fetchTokenFromUserid,
+  fetchTokenFromId,
   deleteUserToken,
 } = require("../../services/UserToken");
 
 const resetPasswordPayloadSchema = Joi.object().keys({
   newPassword: Joi.string().trim().min(8).max(30).required(),
   confirmPassword: Joi.string().trim().min(8).max(30).required(),
-  userId: Joi.string().required(),
+  token: Joi.string().trim().required(),
 });
 
-const resetPasswordController = async (req, res,next) => {
+const resetPasswordController = async (req, res, next) => {
   try {
     const result = req.body;
+    const { resetPasswordSession } = req.cookies;
+
+    if (!resetPasswordSession) {
+      return res.status(401).json({
+        error: "Unauthorized",
+        message: "User not signed in",
+        statusCode: 401,
+      });
+    }
+
+    const decodedData = JWT.verify(
+      resetPasswordSession,
+      process.env.JWT_SECRET
+    );
+    const { userId } = decodedData;
     const { error, value } = resetPasswordPayloadSchema.validate(result);
-    
+
     if (error) {
-      return res.status(402).json({ message: error });
+      return res.status(402).json({ message: error.message });
     }
 
     if (value.confirmPassword === value.newPassword) {
       const hashedPassword = await bcrypt.hash(value.newPassword, 10);
-      const userRef = await fetchUserFromId(value.userId);
+      const userRef = await fetchUserFromId(userId);
       const result = await updatePasswordService(userRef, hashedPassword);
       if (result) {
-        let deleteTokenRef = await fetchTokenFromUserid(value.userId);
-        await deleteUserToken(deleteTokenRef);
-        return res
-          .status(200)
-          .json({ message: "Password updated Successfully" });
+        let deleteTokenRef = await fetchTokenFromId(value.token);
+        if (deleteTokenRef.token === value.token) {
+          await deleteUserToken(deleteTokenRef);
+          return res
+            .status(200)
+            .json({ message: "Password updated Successfully" });
+        }
+        return res.status(403).json({
+          error: http.STATUS_CODES[403],
+          message: "Invalid credentials",
+          statusCode: 403,
+        });
       } else {
         return res.status(400).json({
           error: STATUS_CODES[400],
@@ -43,7 +66,7 @@ const resetPasswordController = async (req, res,next) => {
       }
     }
   } catch (error) {
-    next(err);
+    next(error);
   }
 };
